Refactor: Kafka Producer 서비스 분리
